Add files via upload
This code fixes an issue where the program would not run due to being unable to determine the IP address of the host PC automatically. To get around having to manually input this information, so the program can both be easily shared, and also to avoid changes to locally assigned IP addresses due to DHCP, the code will now check what the local IP address is each time it is run.